requirements.txt
APScheduler==3.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.25.2
Werkzeug==0.15.4
